requirements.txt
google-api-python-client
google-auth
tqdm
requests
colorama